Move OFN::SubscriptionService to SubscriptionVariantsService
