+++ app/serializers/api/admin/subscription_line_item_serializer.rb
@@ -13,9 +13,9 @@ module Api
       end
 
       def in_open_and_upcoming_order_cycles
-        OpenFoodNetwork::SubscriptionService
-          .in_open_and_upcoming_order_cycles?(option_or_assigned_shop, option_or_assigned_schedule,
-                                              object.variant)
+        SubscriptionVariantsService.in_open_and_upcoming_order_cycles?(option_or_assigned_shop,
+                                                                       option_or_assigned_schedule,
+                                                                       object.variant)
       end
 
       private
